Add more question types.
