+++ QuantaTools/__init__.py
@@ -44,9 +44,10 @@ from .maths_tools.maths_utilities import set_maths_vocabulary, set_maths_questio
     tokens_to_answer, insert_question_number, make_a_maths_question_and_answer
 from .maths_tools.maths_complexity import get_maths_question_complexity, get_maths_min_complexity, calc_maths_quanta_for_position_nodes
 from .maths_tools.maths_data_generator import maths_data_generator_core, maths_data_generator, make_maths_questions_and_answers
-from .maths_tools.maths_test_questions import make_maths_test_questions_and_answers, test_maths_questions_by_complexity, \
-    test_maths_questions_by_impact, test_maths_questions_and_add_useful_node_tags, TRICASE_QUESTIONS, \
-    make_maths_tricase_questions, test_correctness_on_num_questions
+from .maths_tools.maths_test_questions import make_maths_test_questions_and_answers
+from .maths_tools.maths_test_questions.test_questions_checker import test_maths_questions_by_complexity, \
+    test_maths_questions_by_impact, test_maths_questions_and_add_useful_node_tags, test_correctness_on_num_questions
+from . import TRICASE_QUESTIONS, make_maths_tricase_questions
 from .maths_tools.maths_algo_search import run_intervention_core, run_strong_intervention, run_weak_intervention, \
     succeed_test, math_common_prereqs, \
     add_ss_tag, add_ss_prereqs, add_ss_test, \
