+++ QuantaTools/maths_tools/maths_test_questions/__init__.py
@@ -1,0 +1,2 @@
+from .manual_test_questions_generator import *
+from .tricase_test_questions_generator import *
